+++ lib/rpc_grpc/simulation.ml
@@ -1,0 +1,1 @@
+type t = { gas_wanted : int64; gas_used : int64 }